Updated ChallengeController to a partial class so new challenges can be added without changing the original file
